Transform Main.buildSplitForest into class SplitIssueForest (#14)
