+++ src/main/scala/me/rexim/issuestant/Main.scala
@@ -11,31 +11,6 @@ import me.rexim.issuestant.github._
 import me.rexim.issuestant.splittree._
 
 object Main {
-  def buildSplitForest(issues: List[Issue]): List[SplitIssueTree] = {
-    val initialForestMap: Map[Int, SplitIssueTree] =
-      issues.map { issue =>
-        issue.number -> SplitIssueTree.fromIssue(issue)
-      }.toMap
-
-    val forestRootNumbers: Set[Int] =
-      issues
-        .filter(_.parentNumber.isEmpty)
-        .map(_.number)
-        .toSet
-
-    issues.foldLeft(initialForestMap) {
-      case (forestMap, issue) =>
-        issue.parentNumber.map { parentNumber =>
-          val Some(parent) = forestMap.get(parentNumber)
-          val Some(child) = forestMap.get(issue.number)
-
-          forestMap.updated(
-            parentNumber,
-            parent.copy(children = child :: parent.children))
-        } getOrElse(forestMap)
-    }.values.filter(tree => forestRootNumbers.contains(tree.number)).toList
-  }
-
   def main(args: Array[String]): Unit = {
     // Circe macros expansions use asInstanceOf, which makes
     // Wartremover freak out a little bit. To avoid that we just
@@ -44,7 +19,6 @@ object Main {
     val issues =
       decode[List[Issue]](Source.fromFile("issues.json").mkString)
 
-    issues.foreach(issues => buildSplitForest(issues).foreach(_.print()))
-
+    issues.foreach(issues => println(new SplitIssueForest(issues)))
   }
 }
